+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## Unreleased
 * Add support for tags when creating new orchestrations ([#231](https://github.com/microsoft/durabletask-java/pull/230))
+* Add versioning support for client and worker ([#224](https://github.com/microsoft/durabletask-java/pull/224))
 
 ## v1.5.2
 * Add distributed tracing support for Azure Functions client scenarios ([#211](https://github.com/microsoft/durabletask-java/pull/211))
-
 
 ## v1.5.1
 * Improve logging for unexpected connection failures in DurableTaskGrpcWorker ([#216](https://github.com/microsoft/durabletask-java/pull/216/files))